libft v0.1β (massif)
collection: virtual methods (2x speedup, yay), opt_put(); zipall(), unzip(), span(), partition(), flatten(), flatmap(), countif(), findif(), indexwhere(), spltted(), distinct() and total rework
libft: ft_printf() updated, ft_strtod(), ft_strtoimax(), ft_qsort(), swap(), readfile(), string utils, mv ctype
tuple, triad updated
bitset updated
utils updated
avx wip
…and slightly normed

+++ src/collection/tree/treeobj.h
@@ -13,7 +13,8 @@
 #ifndef TREEOBJ_H
 # define TREEOBJ_H
 
-# include "collection/abstractobj.h"
+# include "collection/abstract.h"
+# include "collection/internal/object.h"
 
 typedef struct s_node	t_node;
 
@@ -25,7 +26,7 @@ struct	s_node {
 
 struct	s_object_meta {
 	size_t	itemsize;
-	ssize_t	size;
+	long	size;
 	t_node	*root;
 	size_t	deletions;
 };
